requirements.txt
asgiref==3.11.1
colorama==0.4.6
Django==6.0.4
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.3.0
iniconfig==2.3.0
packaging==26.1
pluggy==1.6.0
Pygments==2.20.0
PyJWT==2.12.1
pytest==9.0.3
pytest-django==4.12.0
sqlparse==0.5.5
tzdata==2026.1
whitenoise==6.12.0
